Real data, date changes, Tracked items
Added real data for Milk and Ground Beef 80/20.
Modified dates of mock data to reflect the past and reduced the amount of data to a little over 2 years, instead of 8.
Tracked Items have a message now if they are available below the target price. It currently works with prices based on what the date selected on settings.html.