Savings percent, tracked ingredients, Shopping list prices
Started work on savings percent algorithm, need best mix to finish
Moved the date picker from the settings page to the user ingredients page to make it easier to switch dates.
Added a shopping list total to the shopping list page. The algorithm will have to calculate it, so it currently only works for best store.